models: moves the folder in from `worker` folder
