+++ tsconfig.json
@@ -20,6 +20,7 @@
   ],
   "include": [
     "src/**/*",
-    "worker/**/*"
+    "worker/**/*",
+    "models/**/*"
   ]
 }
